mapa: salida hacia el metro, ford y trolebus
