+++ app/src/main/java/ovh/gabrielhuav/sensores_escom_v2/presentation/components/ipn.zacatenco.escom.mainBuilding/ServerConnectionManager.kt
@@ -11,7 +11,7 @@ class ServerConnectionManager(
     private val context: Context,
     val onlineServerManager: OnlineServerManager
 ) {
-    private val serverUrl = "ws://192.168.1.71:3000"
+    private val serverUrl = "ws://192.168.100.112:3000"
     private var isConnecting = false
     private val mainHandler = Handler(Looper.getMainLooper())
